requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.3
pydantic-settings==2.1.0